+++ Troops/Byzantium/ByzantiumCataphract.cs
@@ -9,11 +9,11 @@ namespace WarSimulator.Troops.Byzantium
 {
     public class ByzantiumCataphract : ICataphract
     {
-        public int Attack { get; set; }
-        public int Defence { get; set; }
-        public int Speed { get; set; }
-        public int Life { get; set; }
-        public int Price { get; set; }
+        public double Attack { get; set; }
+        public double Defence { get; set; }
+        public double Speed { get; set; }
+        public double Life { get; set; }
+        public double Price { get; set; }
         public ByzantiumCataphract()
         {
             Attack = 30;
